fstman: set FST link loss configuration on/off in different scenarios
When a backup interface exists for specific peer, set FST link loss
configuration of the active interface to on.
The idea is once link quality is degraded on active interface, we would
like to switch to backup interface as soon as possible.
When backup interface doesn't exist, set FST link loss configuration
to off, to maintain link up even if link quality is poor.

FST link loss configuration is set using sysfs file.

Change-Id: I16c0d18bf9e855f7996a2e8639f17ac090c40673
Signed-off-by: Hamad Kadmany <hkadmany@codeaurora.org>